The ABA doesn't care about Comatose, either

+++ db/migrate/20090324032935_convert_comatose_pages.rb
@@ -1,6 +1,6 @@
 class ConvertComatosePages < ActiveRecord::Migration
   def self.up
-    return if ASSOCIATION.short_name == "MBRA"
+    return if ASSOCIATION.short_name == "MBRA" || ASSOCIATION.short_name == "ABA"
     
     Page.delete_all
     Page::Version.delete_all
